pyproject.toml
[tool.poetry]
name = "percona-helm-charts"
license = "Apache License, Version 2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.3"
deepdiff = "^7.0.1"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
